Add Bitch, Fenris, GFBlaster, Ns3-Cr.KT, Nyx, UMP, USP
Also took the opportunity to organize/refactor some of the files in the root.  Will definitely want to do this more as I try to bring the weapon pack up to date...

NOTE: The UMP & USP do not have credits files!  I will want to make a proper note in the readme eventually regarding this, I don't want to leave their assets uncredited...