Enhance PD-42
Add some loadout codes to allow setting the initial firemode, as well as whether or not the weapon uses the original reflex sight or the newly supported ironsight variant.
Refactored the gun frames from their firing overlays, which should help clean up assets and make future work a bit easier?
The script refactoring was modeled off of the SMG as the two are surprisingly similar in structure