Adjust burnRange math
At worst, the cone should now be 0.65 * 1.5, which is 0.975, clamped to 1x the HDCONST_ONEMETRE, which is slightly smaller than the flat 1.5x range.
At its furthest, the cone should now be 2.15 * 1.5, which is 3.225 which is double the flat 1.5x range.

In testing, enemies right in front of the player get immolated if they were shot at directly, which makes sense.  A cluster of enemies towards the far end of the cone can all get hit, leaving this feeling much more like a crowd control weapon rather than a short-range beam weapon, much better for a flamethrower.